Migrate common types to common_types.go
