--- pkg/apis/rke.cattle.io/v1/certificaterotation.go
@@ -1,6 +1,0 @@
-package v1
-
-type RotateCertificates struct {
-	Generation int64    `json:"generation,omitempty"`
-	Services   []string `json:"services,omitempty"`
-}
